+++ cjwstate/tests/modules/i18n/catalogs/extract/test_spec.py
@@ -37,7 +37,6 @@ class FindSpecMessagesTest(unittest.TestCase):
                 "has_zen_mode": True,
                 "row_action_menu_entry_title": "Solve your problem",
                 "help_url": "/testme",
-                "parameters_version": 3,
             }
         )
         result = find_spec_messages(spec)